+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Supply Personal Data Disclosure form. Refs STCOM-1449.
 * Re-incorporate component Readme's into storybook. Refs STCOM-1258.
+* Update timezone data. Refs STCOM-1472.
 
 ## [13.1.0](https://github.com/folio-org/stripes-components/tree/v13.1.0) (2026-04-14)